Improve highlighting of curl(1) options list
This commit makes the `.curlrc` grammar tolerant of non-breaking spaces,
which curl(1) doesn't treat as whitespace for the purposes of delimiting
tokens (but regex engines do).

We make this change chiefly to benefit `curlrc-options.tsv`, which lists
short-names before long-names, something that normally causes the latter
to be highlighted as an argument of the former.